refactor: split sim_codegen/{fsm,pipeline}.rs into submodules
sim_codegen.rs had grown to ~7456 LoC. Convert to a module folder and
extract the two largest per-construct emitters into sibling files:

  src/sim_codegen/mod.rs       6644  (was src/sim_codegen.rs)
  src/sim_codegen/fsm.rs        417  (gen_fsm)
  src/sim_codegen/pipeline.rs   424  (gen_pipeline + its helpers)

Each submodule extends impl SimCodegen with a single pub(super) gen_*
entry point and reaches back into the shared free-function helpers via
'use super::*'. No visibility bumps to pub(crate) were needed.

Why these two first: per-construct and self-contained (gen_fsm has no
self.* calls; gen_pipeline's self.* calls are all pipeline-specific
helpers that move with it). Other emitters (gen_ram, gen_fifo, etc.)
can follow the same pattern in follow-up PRs as the need arises.

No behavior change. cargo test: 132/132 pass, unchanged.
