+++ src/sim_codegen/mod.rs
@@ -19,6 +19,12 @@ use crate::comb_graph;
 use crate::resolve::{Symbol, SymbolTable};
 use crate::typecheck::enum_width;
 
+// Per-construct emitters split out to keep this file from growing further.
+// Each submodule extends `impl SimCodegen` with a single `gen_*` entry
+// point and calls back into the shared helpers in this file via `super::`.
+mod fsm;
+mod pipeline;
+
 // ── Public API ────────────────────────────────────────────────────────────────
 
 pub struct SimModel {
@@ -5157,411 +5163,6 @@ impl<'a> SimCodegen<'a> {
 
 // ── FSM codegen ───────────────────────────────────────────────────────────────
 
-impl<'a> SimCodegen<'a> {
-    fn gen_fsm(&self, f: &FsmDecl) -> SimModel {
-        let name = &f.name.name;
-        let class = format!("V{name}");
-        let enum_map = build_enum_map(self.symbols);
-
-        // Collect bus port names and flattened signals (same pattern as gen_module)
-        let mut bus_port_names: HashSet<String> = HashSet::new();
-        let mut bus_flat: Vec<(String, TypeExpr)> = Vec::new();
-        for p in &f.ports {
-            if let Some(ref bi) = p.bus_info {
-                bus_port_names.insert(p.name.name.clone());
-                bus_flat.extend(flatten_bus_port(&p.name.name, bi, self.symbols));
-            }
-        }
-
-        let mut port_names: HashSet<String> = f.ports.iter()
-            .filter(|p| p.bus_info.is_none())
-            .map(|p| p.name.name.clone())
-            .collect();
-        for (flat_name, _) in &bus_flat {
-            port_names.insert(flat_name.clone());
-        }
-
-        let empty_regs  = HashSet::new();
-        let empty_lets  = HashSet::new();
-        let empty_insts = HashSet::new();
-        let empty_wide  = HashSet::new();
-        let empty_w     = HashMap::new();
-
-        let n_states   = f.state_names.len();
-        let state_bits = enum_width(n_states);
-        let state_ty   = cpp_uint(state_bits as u32);
-
-        let state_idx: HashMap<String, usize> = f.state_names.iter()
-            .enumerate().map(|(i, s)| (s.name.clone(), i)).collect();
-        let default_idx = state_idx.get(&f.default_state.name).copied().unwrap_or(0);
-
-        let (rst_name, _is_async, is_low) = extract_reset_info(&f.ports);
-        let rst_cond = if is_low { format!("(!{})", rst_name) } else { rst_name.clone() };
-
-        let mut h = String::new();
-        h.push_str("#pragma once\n#include <cstdint>\n#include <cstdio>\n#include \"verilated.h\"\n\n");
-        // Emit param constants as #define
-        for p in &f.params {
-            if matches!(p.kind, ParamKind::Const | ParamKind::WidthConst(..)) {
-                if let Some(ref def) = p.default {
-                    let val = eval_const_expr(def);
-                    h.push_str(&format!("#ifndef {}\n#define {} {val}ULL\n#endif\n", p.name.name, p.name.name));
-                }
-            }
-        }
-        // Collect FSM Vec port infos (for flat field emission and internal arrays)
-        struct FsmVecPortInfo {
-            name: String,
-            elem_ty: String,
-            count: u64,
-            is_input: bool,
-            is_port_reg: bool,
-        }
-        let fsm_vec_port_infos: Vec<FsmVecPortInfo> = f.ports.iter()
-            .filter_map(|p| {
-                if let Some((elem_ty, count_str)) = vec_array_info(&p.ty) {
-                    let count: u64 = count_str.parse().unwrap_or(0);
-                    Some(FsmVecPortInfo {
-                        name: p.name.name.clone(),
-                        elem_ty,
-                        count,
-                        is_input: p.direction == Direction::In,
-                        is_port_reg: p.reg_info.is_some(),
-                    })
-                } else {
-                    None
-                }
-            })
-            .collect();
-        let fsm_vec_port_names: HashSet<String> = fsm_vec_port_infos.iter().map(|v| v.name.clone()).collect();
-        // All FSM Vec ports have internal C arrays `_name[N]` and always resolve to `_name` in ctx.
-        // (Both input and output Vec ports, whether port-reg or not.)
-        let fsm_vec_port_reg_names: HashSet<String> = fsm_vec_port_names.clone();
-        // Vec-typed regs in f.regs also need array subscript in Index expressions.
-        let fsm_vec_reg_names: HashSet<String> = f.regs.iter()
-            .filter(|r| matches!(r.ty, TypeExpr::Vec(..)))
-            .map(|r| r.name.name.clone())
-            .collect();
-        // All Vec names for the ctx (so Index uses `[i]` syntax): ports + regs
-        let mut fsm_vec_names: HashSet<String> = fsm_vec_port_names.clone();
-        fsm_vec_names.extend(fsm_vec_reg_names.iter().cloned());
-
-        h.push('\n');
-        h.push_str(&format!("class {class} {{\npublic:\n  // State constants\n"));
-        for (i, sn) in f.state_names.iter().enumerate() {
-            h.push_str(&format!("  static const {state_ty} STATE_{} = {i};\n", sn.name.to_uppercase()));
-        }
-        h.push('\n');
-        // Port fields: Vec ports → N flat fields; bus ports → flattened signals; others → single field
-        for p in &f.ports {
-            if bus_port_names.contains(&p.name.name) {
-                continue; // bus ports emitted separately as flattened signals
-            } else if let Some(vi) = fsm_vec_port_infos.iter().find(|v| v.name == p.name.name) {
-                for i in 0..vi.count {
-                    h.push_str(&format!("  {} {}_{i};\n", vi.elem_ty, vi.name));
-                }
-            } else {
-                h.push_str(&format!("  {} {};\n", cpp_port_type(&p.ty), p.name.name));
-            }
-        }
-        // Flattened bus port fields
-        for (flat_name, flat_ty) in &bus_flat {
-            let ty = cpp_port_type(flat_ty);
-            h.push_str(&format!("  {ty} {flat_name};\n"));
-        }
-        // Datapath registers as public members (accessible from testbench)
-        for reg in &f.regs {
-            if let Some((elem_ty, count)) = vec_array_info(&reg.ty) {
-                h.push_str(&format!("  {} {}[{}];\n", elem_ty, reg.name.name, count));
-            } else {
-                let ty = cpp_internal_type(&reg.ty);
-                h.push_str(&format!("  {} {};\n", ty, reg.name.name));
-            }
-        }
-        // Let bindings as public members
-        for lb in &f.lets {
-            let ty = lb.ty.as_ref().map(|t| cpp_internal_type(t)).unwrap_or_else(|| "uint32_t".to_string());
-            h.push_str(&format!("  {} {};\n", ty, lb.name.name));
-        }
-        // Wire declarations as public members
-        for w in &f.wires {
-            let ty = cpp_internal_type(&w.ty);
-            h.push_str(&format!("  {} {};\n", ty, w.name.name));
-        }
-        h.push('\n');
-
-        // Constructor inits: skip Vec/bus port names (they use flat fields), add flat field inits
-        let port_inits: Vec<String> = f.ports.iter()
-            .filter(|p| !fsm_vec_port_names.contains(&p.name.name) && !bus_port_names.contains(&p.name.name))
-            .map(|p| format!("{}(0)", p.name.name))
-            .collect();
-        let vec_port_flat_inits: Vec<String> = fsm_vec_port_infos.iter()
-            .flat_map(|vi| (0..vi.count).map(move |i| format!("{}_{i}(0)", vi.name)))
-            .collect();
-        let bus_flat_inits: Vec<String> = bus_flat.iter()
-            .map(|(n, _)| format!("{n}(0)"))
-            .collect();
-        let reg_inits: Vec<String> = f.regs.iter()
-            .filter(|r| !matches!(r.ty, TypeExpr::Vec(..)))  // Vec regs use memset in ctor body
-            .map(|r| {
-                let init_expr = reset_value_from_reg_reset(&r.reset)
-                    .or(r.init.as_ref());
-                if let Some(expr) = init_expr {
-                    let init_val = cpp_expr(expr, &Ctx::new(&empty_regs, &port_names, &empty_lets, &empty_insts, &empty_wide, &empty_w, &enum_map, &bus_port_names));
-                    format!("{}({})", r.name.name, init_val)
-                } else {
-                    format!("{}(0)", r.name.name)
-                }
-            }).collect();
-        let state_inits = vec!["_clk_prev(0)".to_string(), format!("_state_r({default_idx})")];
-        let all_inits: Vec<String> = port_inits.into_iter()
-            .chain(vec_port_flat_inits)
-            .chain(bus_flat_inits)
-            .chain(reg_inits)
-            .chain(state_inits)
-            .collect();
-        // Constructor body: memset internal arrays for Vec ports + Vec regs
-        let mut fsm_vec_memsets: Vec<String> = fsm_vec_port_infos.iter()
-            .map(|vi| format!("    memset(_{}, 0, sizeof(_{}));", vi.name, vi.name))
-            .collect();
-        // Vec regs in FSM: public array members, initialized via memset
-        for reg in &f.regs {
-            if matches!(reg.ty, TypeExpr::Vec(..)) {
-                let n = &reg.name.name;
-                fsm_vec_memsets.push(format!("    memset({n}, 0, sizeof({n}));"));
-            }
-        }
-        if fsm_vec_memsets.is_empty() {
-            h.push_str(&format!("  {class}() : {} {{}}\n", all_inits.join(", ")));
-        } else {
-            h.push_str(&format!("  {class}() : {} {{\n", all_inits.join(", ")));
-            for ms in &fsm_vec_memsets { h.push_str(&format!("{ms}\n")); }
-            h.push_str("  }\n");
-        }
-        h.push_str(&format!("  explicit {class}(VerilatedContext*) : {class}() {{}}\n"));
-        h.push_str("  void eval();\n  void eval_posedge();\n  void eval_comb();\n  void final() { trace_close(); }\nprivate:\n");
-        // Private internal arrays for Vec ports
-        for vi in &fsm_vec_port_infos {
-            h.push_str(&format!("  {} _{}[{}];\n", vi.elem_ty, vi.name, vi.count));
-        }
-        h.push_str("  uint8_t _clk_prev;\n");
-        h.push_str(&format!("  {state_ty} _state_r;\n"));
-        // };\n deferred until after trace support added
-
-        let mut cpp = String::new();
-        cpp.push_str(&format!("#include \"{class}.h\"\n\n"));
-
-        let clk_port = f.ports.iter().find(|p| matches!(&p.ty, TypeExpr::Clock(_)))
-            .map(|p| p.name.name.as_str()).unwrap_or("clk");
-
-        cpp.push_str(&format!("void {class}::eval() {{\n"));
-        cpp.push_str("  if (!_trace_fp && Verilated::traceFile() && Verilated::claimTrace())\n");
-        cpp.push_str("    trace_open(Verilated::traceFile());\n");
-        cpp.push_str("  eval_comb();\n  eval_posedge();\n  eval_comb();\n");
-        cpp.push_str("  if (_trace_fp) trace_dump(_trace_time++);\n");
-        cpp.push_str("}\n\n");
-
-        let fsm_reg_names: HashSet<String> = f.regs.iter().map(|r| r.name.name.clone()).collect();
-        let fsm_let_names: HashSet<String> = f.lets.iter().map(|l| l.name.name.clone())
-            .chain(f.wires.iter().map(|w| w.name.name.clone())).collect();
-        let mut fsm_widths: HashMap<String, u32> = HashMap::new();
-        for p in &f.ports { fsm_widths.insert(p.name.name.clone(), type_bits_te(&p.ty)); }
-        for r in &f.regs { fsm_widths.insert(r.name.name.clone(), type_bits_te(&r.ty)); }
-        for l in &f.lets {
-            if let Some(ty) = &l.ty { fsm_widths.insert(l.name.name.clone(), type_bits_te(ty)); }
-        }
-        for w in &f.wires { fsm_widths.insert(w.name.name.clone(), type_bits_te(&w.ty)); }
-        let ctx_fsm = {
-            let mut c = Ctx::new(&fsm_reg_names, &port_names, &fsm_let_names, &empty_insts,
-                                 &empty_wide, &fsm_widths, &enum_map, &bus_port_names)
-                .with_vec_names(&fsm_vec_names)
-                .with_fsm_vec_port_regs(&fsm_vec_port_reg_names);
-            c.fsm_mode = true;
-            c
-        };
-
-        cpp.push_str(&format!("void {class}::eval_posedge() {{\n"));
-        cpp.push_str(&format!("  bool _rising = ({clk_port} && !_clk_prev);\n"));
-        cpp.push_str(&format!("  _clk_prev = {clk_port};\n"));
-        cpp.push_str("  if (!_rising) return;\n");
-        cpp.push_str(&format!("  {state_ty} _n_state = _state_r;\n"));
-        // Shadow variables for datapath regs
-        for reg in &f.regs {
-            let n = &reg.name.name;
-            if let Some((elem_ty, count)) = vec_array_info(&reg.ty) {
-                cpp.push_str(&format!("  {elem_ty} _n_{n}[{count}]; memcpy(_n_{n}, {n}, sizeof({n}));\n"));
-            } else {
-                let ty = cpp_internal_type(&reg.ty);
-                cpp.push_str(&format!("  {ty} _n_{n} = {n};\n"));
-            }
-        }
-        cpp.push_str(&format!("  if ({rst_cond}) {{\n    _n_state = {default_idx};\n"));
-        // Reset datapath regs
-        for reg in &f.regs {
-            let reset_expr = reset_value_from_reg_reset(&reg.reset)
-                .or(reg.init.as_ref());
-            if let Some(expr) = reset_expr {
-                let n = &reg.name.name;
-                if vec_array_info(&reg.ty).is_some() {
-                    let init_val = cpp_expr(expr, &ctx_fsm);
-                    let count = if let TypeExpr::Vec(_, c) = &reg.ty { eval_const_expr(c) } else { 0 };
-                    cpp.push_str(&format!("    for (int _i = 0; _i < {count}; _i++) _n_{n}[_i] = {init_val};\n"));
-                } else {
-                    let init_val = cpp_expr(expr, &ctx_fsm);
-                    cpp.push_str(&format!("    _n_{n} = {init_val};\n"));
-                }
-            }
-        }
-        // Reset Vec port-regs
-        for vi in &fsm_vec_port_infos {
-            if vi.is_port_reg {
-                let p = f.ports.iter().find(|p| p.name.name == vi.name).unwrap();
-                let reset_expr = p.reg_info.as_ref().and_then(|ri| reset_value_from_reg_reset(&ri.reset).or(ri.init.as_ref()));
-                let reset_val = if let Some(expr) = reset_expr {
-                    cpp_expr(expr, &ctx_fsm)
-                } else {
-                    "0".to_string()
-                };
-                cpp.push_str(&format!("    for (int _i = 0; _i < {}; _i++) _{}[_i] = {};\n",
-                    vi.count, vi.name, reset_val));
-            }
-        }
-        cpp.push_str("  } else {\n");
-        let ctx_posedge = {
-            let mut c = Ctx::new(&fsm_reg_names, &port_names, &fsm_let_names, &empty_insts,
-                                 &empty_wide, &fsm_widths, &enum_map, &bus_port_names)
-                .with_vec_names(&fsm_vec_names)
-                .with_fsm_vec_port_regs(&fsm_vec_port_reg_names);
-            c.posedge_lhs = true;
-            c.fsm_mode = true;
-            c
-        };
-        // Default sequential assignments
-        for stmt in &f.default_seq {
-            let mut body = String::new();
-            emit_reg_stmt(stmt, &ctx_posedge, &mut body, 2);
-            cpp.push_str(&body);
-        }
-        cpp.push_str("    switch (_state_r) {\n");
-        for sb in &f.states {
-            let idx = state_idx.get(&sb.name.name).copied().unwrap_or(0);
-            cpp.push_str(&format!("      case {idx}: // {}\n", sb.name.name));
-            // Emit seq_stmts for this state
-            for stmt in &sb.seq_stmts {
-                let mut body = String::new();
-                emit_reg_stmt(stmt, &ctx_posedge, &mut body, 4);
-                cpp.push_str(&body);
-            }
-            for tr in &sb.transitions {
-                let cond = cpp_expr(&tr.condition, &ctx_fsm);
-                let target_idx = state_idx.get(&tr.target.name).copied().unwrap_or(0);
-                if self.debug_fsm {
-                    // Escape the condition for printf literal
-                    let cond_escaped = cond.replace('\\', "\\\\").replace('"', "\\\"").replace('%', "%%");
-                    cpp.push_str(&format!(
-                        "        if ({cond}) {{ _n_state = {target_idx}; \
-                         printf(\"[FSM][{name}] {src} -> {tgt} ({cond_lit})\\n\"); break; }}\n",
-                        src = sb.name.name,
-                        tgt = tr.target.name,
-                        cond_lit = cond_escaped,
-                    ));
-                } else {
-                    cpp.push_str(&format!("        if ({cond}) {{ _n_state = {target_idx}; break; }}\n"));
-                }
-            }
-            cpp.push_str("        break;\n");
-        }
-        cpp.push_str("    }\n  }\n  _state_r = _n_state;\n");
-        // Commit datapath regs
-        for reg in &f.regs {
-            let n = &reg.name.name;
-            if vec_array_info(&reg.ty).is_some() {
-                cpp.push_str(&format!("  memcpy({n}, _n_{n}, sizeof({n}));\n"));
-            } else {
-                cpp.push_str(&format!("  {n} = _n_{n};\n"));
-            }
-        }
-        // Fan out Vec port-reg internal arrays to flat public fields
-        for vi in &fsm_vec_port_infos {
-            if vi.is_port_reg {
-                for i in 0..vi.count {
-                    cpp.push_str(&format!("  {}_{i} = _{}[{i}];\n", vi.name, vi.name));
-                }
-            }
-        }
-        cpp.push_str("}\n\n");
-
-        cpp.push_str(&format!("void {class}::eval_comb() {{\n"));
-        // Flat → internal bridge for input Vec ports
-        for vi in &fsm_vec_port_infos {
-            if vi.is_input && !vi.is_port_reg {
-                for i in 0..vi.count {
-                    cpp.push_str(&format!("  _{}[{i}] = {}_{i};\n", vi.name, vi.name));
-                }
-            }
-        }
-        // Let bindings
-        for lb in &f.lets {
-            let val = cpp_expr(&lb.value, &ctx_fsm);
-            cpp.push_str(&format!("  {} = {};\n", lb.name.name, val));
-        }
-        // Default combinational assignments
-        {
-            let mut body = String::new();
-            emit_comb_stmts(&f.default_comb, &ctx_fsm, &mut body, 1);
-            cpp.push_str(&body);
-        }
-        cpp.push_str("  switch (_state_r) {\n");
-        for sb in &f.states {
-            let idx = state_idx.get(&sb.name.name).copied().unwrap_or(0);
-            cpp.push_str(&format!("    case {idx}: {{ // {}\n", sb.name.name));
-            let mut body = String::new();
-            emit_comb_stmts(&sb.comb_stmts, &ctx_fsm, &mut body, 3);
-            cpp.push_str(&body);
-            cpp.push_str("      break;\n    }\n");
-        }
-        cpp.push_str("  }\n");
-        // Internal → flat bridge for output Vec ports (non-reg)
-        for vi in &fsm_vec_port_infos {
-            if !vi.is_input && !vi.is_port_reg {
-                for i in 0..vi.count {
-                    cpp.push_str(&format!("  {}_{i} = _{}[{i}];\n", vi.name, vi.name));
-                }
-            }
-        }
-        cpp.push_str("}\n");
-
-        // Add trace support
-        // Build flat Vec port trace signals (name_i → field name_i, width = elem_width)
-        let mut fsm_flat_vec_traces: Vec<(String, String, u32)> = Vec::new();
-        for vi in &fsm_vec_port_infos {
-            let elem_bits = if let TypeExpr::Vec(elem, _) = f.ports.iter()
-                .find(|p| p.name.name == vi.name).map(|p| &p.ty).unwrap() {
-                type_width(elem)
-            } else { 32 };
-            for i in 0..vi.count {
-                let fname = format!("{}_{i}", vi.name);
-                fsm_flat_vec_traces.push((fname.clone(), fname, elem_bits));
-            }
-        }
-        let mut extra_sigs_owned: Vec<(String, String, u32)> = vec![
-            ("state_r".to_string(), "_state_r".to_string(), state_bits as u32),
-        ];
-        extra_sigs_owned.extend(fsm_flat_vec_traces);
-        // Add flattened bus port signals to trace
-        for (flat_name, flat_ty) in &bus_flat {
-            let bits = type_bits_te(flat_ty);
-            extra_sigs_owned.push((flat_name.clone(), flat_name.clone(), bits));
-        }
-        let extra_sigs_ref: Vec<(&str, &str, u32)> = extra_sigs_owned.iter()
-            .map(|(n, e, w)| (n.as_str(), e.as_str(), *w))
-            .collect();
-        add_trace_to_simple_construct(&mut h, &mut cpp, &class, name, &f.ports, &extra_sigs_ref);
-        h.push_str("};\n");
-
-        SimModel { class_name: class, header: h, impl_: cpp }
-    }
-}
 
 // ── Regfile codegen ───────────────────────────────────────────────────────────
 
@@ -7040,417 +6641,4 @@ impl<'a> SimCodegen<'a> {
         SimModel { class_name: class, header: h, impl_: cpp }
     }
 
-    // ── Pipeline ─────────────────────────────────────────────────────────────
-
-    fn gen_pipeline(&self, p: &PipelineDecl) -> SimModel {
-        let name = &p.name.name;
-        let class = format!("V{name}");
-        let _enum_map = build_enum_map(self.symbols);
-
-        let port_names: HashSet<String> = p.ports.iter()
-            .map(|pt| pt.name.name.clone()).collect();
-        let stage_names: Vec<String> = p.stages.iter()
-            .map(|s| s.name.name.clone()).collect();
-        let stage_prefixes: Vec<String> = stage_names.iter()
-            .map(|s| s.to_lowercase()).collect();
-
-        // Flatten stage registers and let bindings
-        struct StageReg { prefixed: String, ty: String, reset_val: String, bits: u32, is_let: bool }
-        let mut all_regs: Vec<StageReg> = Vec::new();
-        let mut stage_reg_names: Vec<HashSet<String>> = Vec::new();
-
-        for (si, stage) in p.stages.iter().enumerate() {
-            let prefix = &stage_prefixes[si];
-            let mut names_set = HashSet::new();
-            for item in &stage.body {
-                match item {
-                    ModuleBodyItem::RegDecl(r) => {
-                        let prefixed = format!("{}_{}", prefix, r.name.name);
-                        let ty = cpp_internal_type(&r.ty);
-                        let bits = type_bits_te(&r.ty);
-                        let reset_val = Self::pipeline_reset_value(&r.reset).unwrap_or("0".to_string());
-                        names_set.insert(r.name.name.clone());
-                        all_regs.push(StageReg { prefixed, ty, reset_val, bits, is_let: false });
-                    }
-                    ModuleBodyItem::LetBinding(l) => {
-                        // ty=None means assignment to existing port/wire — skip as new binding
-                        if l.ty.is_none() { continue; }
-                        let prefixed = format!("{}_{}", prefix, l.name.name);
-                        let ty = if let Some(ref te) = l.ty { cpp_internal_type(te) } else { "uint32_t".to_string() };
-                        let bits = if let Some(ref te) = l.ty { type_bits_te(te) } else { 32 };
-                        names_set.insert(l.name.name.clone());
-                        all_regs.push(StageReg { prefixed, ty, reset_val: String::new(), bits, is_let: true });
-                    }
-                    _ => {}
-                }
-            }
-            stage_reg_names.push(names_set);
-        }
-
-        let mut reg_names: HashSet<String> = HashSet::new();
-        let mut let_names: HashSet<String> = HashSet::new();
-        let mut widths: HashMap<String, u32> = HashMap::new();
-        for sr in &all_regs {
-            if sr.is_let { let_names.insert(sr.prefixed.clone()); }
-            else { reg_names.insert(sr.prefixed.clone()); }
-            widths.insert(sr.prefixed.clone(), sr.bits);
-        }
-        for prefix in &stage_prefixes {
-            reg_names.insert(format!("{}_valid_r", prefix));
-            widths.insert(format!("{}_valid_r", prefix), 1);
-        }
-        // Add port widths
-        for pt in &p.ports {
-            widths.insert(pt.name.name.clone(), type_bits_te(&pt.ty));
-        }
-
-        let (rst_name, _is_async, is_low) = extract_reset_info(&p.ports);
-        let rst_cond = if is_low { format!("(!{})", rst_name) } else { rst_name.clone() };
-        let clk_name = p.ports.iter()
-            .find(|pt| matches!(pt.ty, TypeExpr::Clock(_)))
-            .map(|pt| pt.name.name.clone())
-            .unwrap_or("clk".to_string());
-
-        // ── Header ──
-        let mut h = String::new();
-        h.push_str("#pragma once\n#include <cstdint>\n#include <cstdio>\n#include \"verilated.h\"\n\n");
-        for param in &p.params {
-            if matches!(param.kind, ParamKind::Const | ParamKind::WidthConst(..)) {
-                if let Some(ref def) = param.default {
-                    let val = eval_const_expr(def);
-                    h.push_str(&format!("#ifndef {}\n#define {} {val}ULL\n#endif\n", param.name.name, param.name.name));
-                }
-            }
-        }
-        h.push_str(&format!("\nclass {class} {{\npublic:\n"));
-        for pt in &p.ports {
-            h.push_str(&format!("  {} {};\n", cpp_port_type(&pt.ty), pt.name.name));
-        }
-        h.push('\n');
-
-        // Constructor
-        let mut inits: Vec<String> = p.ports.iter().map(|pt| format!("{}(0)", pt.name.name)).collect();
-        inits.push("_clk_prev(0)".to_string());
-        for sr in &all_regs { if !sr.is_let { inits.push(format!("_{}(0)", sr.prefixed)); } }
-        for prefix in &stage_prefixes { inits.push(format!("_{}_valid_r(0)", prefix)); }
-        h.push_str(&format!("  {class}() : {} {{}}\n\n", inits.join(", ")));
-
-        h.push_str("  void eval();\n  void eval_posedge();\n  void eval_comb();\n");
-        h.push_str("  void _do_reset();\n  void final_() {}\n");
-        h.push_str("  void trace_open(const char*) {}\n  void trace_dump(uint64_t) {}\n  void trace_close() {}\n\n");
-
-        h.push_str("private:\n");
-        h.push_str("  uint8_t _clk_prev;\n");
-        for sr in &all_regs { if !sr.is_let { h.push_str(&format!("  {} _{};\n", sr.ty, sr.prefixed)); } }
-        for prefix in &stage_prefixes { h.push_str(&format!("  uint8_t _{}_valid_r;\n", prefix)); }
-        h.push_str("};\n");
-
-        // ── Implementation ──
-        let mut cpp = String::new();
-        cpp.push_str(&format!("#include \"V{name}.h\"\n\n"));
-        cpp.push_str(&format!("void {class}::eval() {{ eval_comb(); eval_posedge(); eval_comb(); }}\n\n"));
-
-        // reset()
-        cpp.push_str(&format!("void {class}::_do_reset() {{\n"));
-        for sr in &all_regs {
-            if !sr.is_let {
-                let v = match sr.reset_val.as_str() { "false" | "1'b0" => "0", "true" | "1'b1" => "1", other => other };
-                cpp.push_str(&format!("  _{} = {v};\n", sr.prefixed));
-            }
-        }
-        for prefix in &stage_prefixes { cpp.push_str(&format!("  _{}_valid_r = 0;\n", prefix)); }
-        cpp.push_str("}\n\n");
-
-        // eval_posedge()
-        cpp.push_str(&format!("void {class}::eval_posedge() {{\n"));
-        cpp.push_str(&format!("  bool _rising = ({clk_name} && !_clk_prev);\n  _clk_prev = {clk_name};\n"));
-        cpp.push_str("  if (_rising) {\n");
-        cpp.push_str(&format!("    if ({rst_cond}) {{ _do_reset(); }} else {{\n"));
-
-        // Evaluate let bindings first (they are combinational and may be referenced in seq blocks)
-        for (si, stage) in p.stages.iter().enumerate() {
-            let prefix = &stage_prefixes[si];
-            for item in &stage.body {
-                if let ModuleBodyItem::LetBinding(l) = item {
-                    let val = self.pipeline_sim_expr(&l.value, prefix, si,
-                        &stage_names, &stage_prefixes, &stage_reg_names,
-                        &port_names, &reg_names, &let_names, &widths, &_enum_map);
-                    let ty = if let Some(ref te) = l.ty { cpp_internal_type(te) } else { "uint32_t".to_string() };
-                    let bits = if let Some(ref te) = l.ty { type_bits_te(te) } else { 32 };
-                    let mask = if bits > 0 && bits < 64 { format!(" & 0x{:X}ULL", (1u64 << bits) - 1) } else { String::new() };
-                    cpp.push_str(&format!("      {ty} {}_{} = ({val}){mask};\n", prefix, l.name.name));
-                }
-            }
-        }
-
-        // Process stages in reverse order so later stages read old values
-        // (mimics SV non-blocking assignment semantics)
-        for si in (0..p.stages.len()).rev() {
-            let stage = &p.stages[si];
-            let prefix = &stage_prefixes[si];
-            if si == 0 {
-                cpp.push_str(&format!("      _{}_valid_r = 1;\n", prefix));
-            } else {
-                cpp.push_str(&format!("      _{}_valid_r = _{}_valid_r;\n", prefix, stage_prefixes[si-1]));
-            }
-            for item in &stage.body {
-                if let ModuleBodyItem::RegBlock(rb) = item {
-                    for stmt in &rb.stmts {
-                        self.emit_pipeline_sim_stmt(&mut cpp, stmt, prefix, si,
-                            &stage_names, &stage_prefixes, &stage_reg_names,
-                            &port_names, &reg_names, &let_names, &widths, &_enum_map, 6);
-                    }
-                }
-            }
-        }
-        for flush in &p.flush_directives {
-            let target = flush.target_stage.name.to_lowercase();
-            let cond = self.pipeline_sim_expr(&flush.condition, "", 0,
-                &stage_names, &stage_prefixes, &stage_reg_names,
-                &port_names, &reg_names, &let_names, &widths, &_enum_map);
-            cpp.push_str(&format!("      if ({cond}) {{ _{target}_valid_r = 0; }}\n"));
-        }
-        cpp.push_str("    }\n  }\n}\n\n");
-
-        // eval_comb()
-        cpp.push_str(&format!("void {class}::eval_comb() {{\n"));
-        for (si, stage) in p.stages.iter().enumerate() {
-            let prefix = &stage_prefixes[si];
-            for item in &stage.body {
-                match item {
-                    ModuleBodyItem::LetBinding(l) => {
-                        // ty=None means assignment to existing port/wire — skip as new binding
-                        if l.ty.is_none() { continue; }
-                        let val = self.pipeline_sim_expr(&l.value, prefix, si,
-                            &stage_names, &stage_prefixes, &stage_reg_names,
-                            &port_names, &reg_names, &let_names, &widths, &_enum_map);
-                        let ty = if let Some(ref te) = l.ty { cpp_internal_type(te) } else { "uint32_t".to_string() };
-                        let bits = if let Some(ref te) = l.ty { type_bits_te(te) } else { 32 };
-                        let mask = if bits > 0 && bits < 64 { format!(" & 0x{:X}ULL", (1u64 << bits) - 1) } else { String::new() };
-                        cpp.push_str(&format!("  {ty} {}_{} = ({val}){mask};\n", prefix, l.name.name));
-                    }
-                    ModuleBodyItem::CombBlock(cb) => {
-                        for stmt in &cb.stmts {
-                            self.emit_pipeline_sim_comb_stmt(&mut cpp, stmt, prefix, si,
-                                &stage_names, &stage_prefixes, &stage_reg_names,
-                                &port_names, &reg_names, &let_names, &widths, &_enum_map, 2);
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-        cpp.push_str("}\n");
-
-        SimModel { class_name: class, header: h, impl_: cpp }
-    }
-
-    fn emit_pipeline_sim_stmt(
-        &self, cpp: &mut String, stmt: &Stmt, prefix: &str, si: usize,
-        sn: &[String], sp: &[String], srn: &[HashSet<String>],
-        pn: &HashSet<String>, rn: &HashSet<String>, ln: &HashSet<String>,
-        w: &HashMap<String, u32>, em: &HashMap<String, Vec<(String, u64)>>, indent: usize,
-    ) {
-        let pad = " ".repeat(indent);
-        match stmt {
-            Stmt::Assign(a) => {
-                let tgt = if let ExprKind::Ident(n) = &a.target.kind {
-                    if pn.contains(n) { n.clone() } else { format!("_{}_{}", prefix, n) }
-                } else { self.pipeline_sim_expr(&a.target, prefix, si, sn, sp, srn, pn, rn, ln, w, em) };
-                let val = self.pipeline_sim_expr(&a.value, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                let tgt_key = if let ExprKind::Ident(n) = &a.target.kind { format!("{}_{}", prefix, n) } else { String::new() };
-                let mask = w.get(&tgt_key).and_then(|&b| if b > 0 && b < 64 { Some(format!(" & 0x{:X}ULL", (1u64 << b) - 1)) } else { None }).unwrap_or_default();
-                cpp.push_str(&format!("{pad}{tgt} = ({val}){mask};\n"));
-            }
-            Stmt::IfElse(ie) => {
-                let cond = self.pipeline_sim_expr(&ie.cond, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                cpp.push_str(&format!("{pad}if ({cond}) {{\n"));
-                for s in &ie.then_stmts { self.emit_pipeline_sim_stmt(cpp, s, prefix, si, sn, sp, srn, pn, rn, ln, w, em, indent+2); }
-                if !ie.else_stmts.is_empty() {
-                    cpp.push_str(&format!("{pad}}} else {{\n"));
-                    for s in &ie.else_stmts { self.emit_pipeline_sim_stmt(cpp, s, prefix, si, sn, sp, srn, pn, rn, ln, w, em, indent+2); }
-                }
-                cpp.push_str(&format!("{pad}}}\n"));
-            }
-            Stmt::For(f) => {
-                if let ForRange::Range(ref lo_expr, ref hi_expr) = f.range {
-                    let lo = self.pipeline_sim_expr(lo_expr, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                    let hi = self.pipeline_sim_expr(hi_expr, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                    cpp.push_str(&format!("{pad}for (int {v} = {lo}; {v} <= {hi}; {v}++) {{\n", v=f.var.name));
-                    for s in &f.body { self.emit_pipeline_sim_stmt(cpp, s, prefix, si, sn, sp, srn, pn, rn, ln, w, em, indent+2); }
-                    cpp.push_str(&format!("{pad}}}\n"));
-                }
-            }
-            _ => {}
-        }
-    }
-
-    fn emit_pipeline_sim_comb_stmt(
-        &self, cpp: &mut String, stmt: &CombStmt, prefix: &str, si: usize,
-        sn: &[String], sp: &[String], srn: &[HashSet<String>],
-        pn: &HashSet<String>, rn: &HashSet<String>, ln: &HashSet<String>,
-        w: &HashMap<String, u32>, em: &HashMap<String, Vec<(String, u64)>>, indent: usize,
-    ) {
-        let pad = " ".repeat(indent);
-        match stmt {
-            CombStmt::Assign(a) => {
-                let tgt = if let ExprKind::Ident(n) = &a.target.kind {
-                    if pn.contains(n) { n.clone() } else { format!("{}_{}", prefix, n) }
-                } else { self.pipeline_sim_expr(&a.target, prefix, si, sn, sp, srn, pn, rn, ln, w, em) };
-                let val = self.pipeline_sim_expr(&a.value, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                cpp.push_str(&format!("{pad}{tgt} = {val};\n"));
-            }
-            CombStmt::IfElse(ie) => {
-                let cond = self.pipeline_sim_expr(&ie.cond, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                cpp.push_str(&format!("{pad}if ({cond}) {{\n"));
-                for s in &ie.then_stmts { self.emit_pipeline_sim_comb_stmt(cpp, s, prefix, si, sn, sp, srn, pn, rn, ln, w, em, indent+2); }
-                if !ie.else_stmts.is_empty() {
-                    cpp.push_str(&format!("{pad}}} else {{\n"));
-                    for s in &ie.else_stmts { self.emit_pipeline_sim_comb_stmt(cpp, s, prefix, si, sn, sp, srn, pn, rn, ln, w, em, indent+2); }
-                }
-                cpp.push_str(&format!("{pad}}}\n"));
-            }
-            _ => {}
-        }
-    }
-
-    fn pipeline_sim_expr(
-        &self, expr: &Expr, prefix: &str, si: usize,
-        sn: &[String], sp: &[String], srn: &[HashSet<String>],
-        pn: &HashSet<String>, rn: &HashSet<String>, ln: &HashSet<String>,
-        w: &HashMap<String, u32>, em: &HashMap<String, Vec<(String, u64)>>,
-    ) -> String {
-        let empty = HashSet::new();
-        let empty_rl: HashMap<String, ResetLevel> = HashMap::new();
-        let ctx = Ctx { reg_names: rn, port_names: pn, let_names: ln, inst_names: &empty, wide_names: &empty, widths: w, posedge_lhs: false, fsm_mode: false, enum_map: em, bus_ports: &empty, reset_levels: &empty_rl, vec_names: None, vec_sizes: None, fsm_vec_port_regs: None, ident_subst: None };
-        match &expr.kind {
-            ExprKind::FieldAccess(base, field) => {
-                if let ExprKind::Ident(bn) = &base.kind {
-                    if let Some(idx) = sn.iter().position(|s| s == bn) {
-                        let p = &sp[idx];
-                        let prefixed = format!("{}_{}", p, field.name);
-                        if rn.contains(&prefixed) { return format!("_{}", prefixed); }
-                        if ln.contains(&prefixed) { return prefixed; }
-                        if field.name == "valid_r" { return format!("_{}_valid_r", p); }
-                        return format!("_{}", prefixed);
-                    }
-                }
-                cpp_expr(expr, &ctx)
-            }
-            ExprKind::Ident(name) => {
-                if pn.contains(name) { return name.clone(); }
-                if name == "valid_r" && !prefix.is_empty() { return format!("_{}_valid_r", prefix); }
-                if si < srn.len() && srn[si].contains(name) {
-                    let prefixed = format!("{}_{}", prefix, name);
-                    if rn.contains(&prefixed) { return format!("_{}", prefixed); }
-                    if ln.contains(&prefixed) { return prefixed; }
-                }
-                cpp_expr(expr, &ctx)
-            }
-            ExprKind::Concat(parts) => {
-                let mut total_bits: u32 = 0;
-                let part_widths: Vec<u32> = parts.iter().map(|p2| {
-                    let pw = self.pipeline_sim_expr_width(p2, prefix, si, srn, w, pn);
-                    total_bits += pw;
-                    pw
-                }).collect();
-                let _ = total_bits; // used implicitly
-                let mut result = String::from("(");
-                let mut bit_pos: u32 = part_widths.iter().sum();
-                for (i, part) in parts.iter().enumerate() {
-                    bit_pos -= part_widths[i];
-                    let val = self.pipeline_sim_expr(part, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                    if i > 0 { result.push_str(" | "); }
-                    if bit_pos > 0 { result.push_str(&format!("((uint64_t)({val}) << {bit_pos})")); }
-                    else { result.push_str(&format!("(uint64_t)({val})")); }
-                }
-                result.push(')');
-                result
-            }
-            ExprKind::Binary(op, lhs, rhs) => {
-                let l = self.pipeline_sim_expr(lhs, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                let r = self.pipeline_sim_expr(rhs, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                if *op == BinOp::Implies {
-                    return format!("(!{l} || {r})");
-                }
-                let os = match op {
-                    BinOp::Add | BinOp::AddWrap => "+", BinOp::Sub | BinOp::SubWrap => "-",
-                    BinOp::Mul | BinOp::MulWrap => "*", BinOp::Div => "/", BinOp::Mod => "%",
-                    BinOp::Eq => "==", BinOp::Neq => "!=", BinOp::Lt => "<", BinOp::Gt => ">",
-                    BinOp::Lte => "<=", BinOp::Gte => ">=", BinOp::And => "&&", BinOp::Or => "||",
-                    BinOp::BitAnd => "&", BinOp::BitOr => "|", BinOp::BitXor => "^", BinOp::Shl => "<<", BinOp::Shr => ">>",
-                    BinOp::Implies => unreachable!(),
-                };
-                format!("({l} {os} {r})")
-            }
-            ExprKind::Unary(op, inner) => {
-                let v = self.pipeline_sim_expr(inner, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                match op { UnaryOp::Not => format!("(!{v})"), UnaryOp::BitNot => format!("(~{v})"), UnaryOp::Neg => format!("(-{v})"),
-                    UnaryOp::RedAnd | UnaryOp::RedOr | UnaryOp::RedXor => format!("({v})") }
-            }
-            ExprKind::MethodCall(base, method, args) => {
-                let b = self.pipeline_sim_expr(base, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                match method.name.as_str() {
-                    "trunc" => { if let Some(wa) = args.first() { let bits = eval_const_expr(wa); if bits < 64 { format!("({b} & 0x{:X}ULL)", (1u64 << bits) - 1) } else { b } } else { b } }
-                    "zext" => { format!("(uint64_t)({b})") }
-                    "sext" => { b } // TODO: proper sign extension
-                    _ => b
-                }
-            }
-            ExprKind::BitSlice(base, hi, lo) => {
-                let b = self.pipeline_sim_expr(base, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                let hv = eval_const_expr(hi); let lv = eval_const_expr(lo); let width = hv - lv + 1;
-                if width < 64 { format!("(({b} >> {lv}) & 0x{:X}ULL)", (1u64 << width) - 1) } else { format!("({b} >> {lv})") }
-            }
-            ExprKind::Index(base, idx) => {
-                let b = self.pipeline_sim_expr(base, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                let i = self.pipeline_sim_expr(idx, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                format!("(({b} >> {i}) & 1)")
-            }
-            ExprKind::Literal(lit) => match lit { LitKind::Dec(v) => format!("{v}"), LitKind::Hex(v) | LitKind::Bin(v) => format!("0x{v:X}"), LitKind::Sized(_, v) => format!("{v}") },
-            ExprKind::Bool(b) => if *b { "1".to_string() } else { "0".to_string() },
-            ExprKind::Ternary(c, t, e) => {
-                let cv = self.pipeline_sim_expr(c, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                let tv = self.pipeline_sim_expr(t, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                let ev = self.pipeline_sim_expr(e, prefix, si, sn, sp, srn, pn, rn, ln, w, em);
-                format!("(({cv}) ? ({tv}) : ({ev}))")
-            }
-            ExprKind::Signed(inner) | ExprKind::Unsigned(inner) | ExprKind::Cast(inner, _) => {
-                self.pipeline_sim_expr(inner, prefix, si, sn, sp, srn, pn, rn, ln, w, em)
-            }
-            ExprKind::Clog2(arg) => { let a = self.pipeline_sim_expr(arg, prefix, si, sn, sp, srn, pn, rn, ln, w, em); format!("_arch_clog2({a})") }
-            _ => cpp_expr(expr, &ctx),
-        }
-    }
-
-    fn pipeline_sim_expr_width(&self, expr: &Expr, prefix: &str, si: usize, srn: &[HashSet<String>], w: &HashMap<String, u32>, pn: &HashSet<String>) -> u32 {
-        match &expr.kind {
-            ExprKind::Ident(name) => {
-                if pn.contains(name) { return *w.get(name).unwrap_or(&8); }
-                if si < srn.len() && srn[si].contains(name) { return *w.get(&format!("{}_{}", prefix, name)).unwrap_or(&8); }
-                *w.get(name).unwrap_or(&8)
-            }
-            ExprKind::FieldAccess(base, field) => {
-                if let ExprKind::Ident(bn) = &base.kind { *w.get(&format!("{}_{}", bn.to_lowercase(), field.name)).unwrap_or(&8) } else { 8 }
-            }
-            ExprKind::MethodCall(_, method, args) => match method.name.as_str() { "trunc"|"zext"|"sext"|"resize" => args.first().map(|a| eval_const_expr(a) as u32).unwrap_or(8), _ => 8 },
-            ExprKind::BitSlice(_, hi, lo) => { let h = eval_const_expr(hi); let l = eval_const_expr(lo); (h - l + 1) as u32 }
-            ExprKind::Literal(LitKind::Sized(ww, _)) => *ww,
-            _ => 8,
-        }
-    }
-
-    fn pipeline_reset_value(reset: &RegReset) -> Option<String> {
-        match reset {
-            RegReset::Explicit(_, _, _, val) | RegReset::Inherit(_, val) => {
-                match &val.kind {
-                    ExprKind::Literal(LitKind::Dec(v)) => Some(format!("{v}")),
-                    ExprKind::Literal(LitKind::Hex(v)) => Some(format!("0x{v:X}")),
-                    ExprKind::Bool(b) => Some(if *b { "1".to_string() } else { "0".to_string() }),
-                    _ => Some("0".to_string()),
-                }
-            }
-            _ => Some("0".to_string()),
-        }
-    }
 }
